Implement the "noop" mode.
Adds a "noop" mode where the module acts like it is a fully-operational
RedisGears v2 (triggers and functions) module, while in fact does
nothing. It still loads the RDB, has the same commands, but the commands
do not perform any action, though still can be invoked. The
configuration options remain as well.
