+++ redisgears_core/src/full/function_del_command.rs
@@ -8,7 +8,8 @@ use redis_module::{Context, RedisError, RedisResult, RedisValue, ThreadSafeConte
 use std::iter::Skip;
 use std::vec::IntoIter;
 
-use crate::{get_libraries, Deserialize, Serialize};
+use super::get_libraries;
+use serde::{Deserialize, Serialize};
 
 use mr_derive::BaseObject;
