+++ redisgears_core/src/full/module_full.rs
@@ -1,0 +1,1 @@
+
